+++ src/main/java/xyz/nikitacartes/easyauth/config/ExtendedConfigV1.java
@@ -172,6 +172,11 @@ public class ExtendedConfigV1 extends ConfigTemplate {
             Prevent offline players from joining the server using online usernames.""")
     public boolean preventOfflinePlayersWithOnlineUsernames = false;
 
+    @Comment("""
+            
+            Check offline players with online usernames every time they join the server for online account.""")
+    public boolean checkOfflinePlayersWithOnlineUsernames = false;
+
     public ExtendedConfigV1() {
         super("extended.conf", """
                 ##                          ##